requirements.txt
cx_Freeze==7.1.1
cx_Logging==3.2.0
dearpygui==1.11.1
et-xmlfile==1.1.0
html2text==2024.2.26
lief==0.14.1
Markdown==3.6
markdown-it-py==3.0.0
mdurl==0.1.2
numpy==2.0.0
openpyxl==3.1.5
pandas==2.2.2
pillow==10.4.0
Pygments==2.18.0
python-dateutil==2.9.0.post0
pytz==2024.1
rich==13.7.1
setuptools==70.2.0
six==1.16.0
tzdata==2024.1
wheel==0.43.0
